docs/docker-to-nomad/requirements.txt
PyYAML
jinja2
streamlit
anthropic
ollama
